Bug fix for computer entering sleep which kills zbeacon
The UDP beacon or zbeacon can fail with OSError: [Errno50] if the
computer has entered sleep. According to
https://www.ibm.com/support/knowledgecenter/en/SSLTBW_2.4.0/com.ibm.zos.v2r4.hala001/syserret.htm
the recommended action when encountering a network down error is to try
again, not exit. The code prior to this fix treated all errors on the
UDP socket send call as fatal, including network down, which can be
correctly handled by ignoring this error. This has been tested and now
the beacon continues sending after the computer wakes from sleep. Tested
on MacOS 10.14.5