Another bug fix for sleeping computers, OSError: [errno 51]
In this fix, it was found that a computer that was sleeping could cause
an error on a socket call, ENETUNREACH, or errno 51. This was
unanticipated since the call was on a udp socket that is connectionless
and this resource ->
https://www.ibm.com/support/knowledgecenter/en/SSLTBW_2.4.0/com.ibm.zos.v2r4.hala001/syserret.htm
suggests that this error happens on a call to a socket to connect.
However, the error was the problem of pyre not being able to resume from
wake. So in addition to handling the case of errno 50, network down, we
now handle the case of the socket call failing with network unreachable
as well.