Split version-bumping and deployment into separate workflows
Due to security restrictions, pull requests from forked repositories do not
have access to the workflow's secret. As a result, the "Deploy to PyPI" workflow will fail
and not trigger a deployment (see https://github.com/zeromq/pyre/pull/157#issuecomment-862613841).
In these cases, the deployment needs to be triggered again by manually dispatching the workflow.

To allow for further flexibility, this PR splits the version-bumping from the deployment workflow.
Specifically, this causes the deployment workflow to trigger on all pushed tags, not only those
created by the "Bump version" workflow.